Introduce a metrics registry promise
This introduces a "promise" of a metrics registry, which lazily
registers new metrics, only filling in an actual implementation for the
metrics when an actual registry is provided.  This simplifies the
creation of new metrics in client-go, as well as the use of metrics in
consuming packages.

A prometheus based implementation is provide in a sub-package, to make
the prometheus dependency optional.